agents/sector_news_analysis/requirements.txt
typer==0.12.5
rich==13.7.1
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.2
tabulate==0.9.0
openai>=1.40.0
click==8.1.7
